chore : Idea Entity 패키지 구조 변경
